+++ src/onegov.org/onegov/org/forms/event.py
@@ -1,3 +1,6 @@
+import json
+
+from cached_property import cached_property
 from datetime import date, datetime, timedelta
 from dateutil import rrule
 from dateutil.rrule import rrulestr
@@ -6,7 +9,7 @@ from onegov.form.fields import MultiCheckboxField
 from onegov.gis import CoordinatesField
 from onegov.org import _
 from sedate import replace_timezone, to_timezone
-from wtforms import StringField, TextAreaField, validators
+from wtforms import RadioField, StringField, TextAreaField, validators
 from wtforms.fields.html5 import DateField, EmailField
 from wtforms_components import TimeField
 
@@ -45,6 +48,8 @@ WEEKDAYS = (
 
 class EventForm(Form):
     """ Defines the form for all events. """
+
+    timezone = 'Europe/Zurich'
 
     email = EmailField(
         label=_("E-Mail"),
@@ -105,16 +110,45 @@ class EventForm(Form):
         validators=[validators.InputRequired()]
     )
 
+    repeat = RadioField(
+        label=_("Repeat"),
+        default='none',
+        choices=(
+            ('without', _("Without")),
+            ('weekly', _("Weekly")),
+            ('dates', _("On additional dates"))
+        )
+    )
+
     weekly = MultiCheckboxField(
         label=_("Repeats itself every"),
         choices=WEEKDAYS,
-        render_kw={'prefix_label': False, 'class_': 'oneline-checkboxes'}
+        render_kw={'prefix_label': False, 'class_': 'oneline-checkboxes'},
+        depends_on=('repeat', 'weekly')
     )
 
     end_date = DateField(
         label=_("Until date"),
-        validators=[validators.Optional()]
+        validators=[validators.Optional()],
+        depends_on=('repeat', 'weekly')
     )
+
+    dates = TextAreaField(
+        label=_("Dates"),
+        depends_on=('repeat', 'dates'),
+        render_kw={'class_': 'many many-dates'},
+    )
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.date_errors = {}
+
+    def on_request(self):
+        self.request.include('common')
+        self.request.include('many')
+
+        if not self.dates.data:
+            self.dates.data = self.dates_to_json(None)
 
     def validate(self):
         """ Make sure a valid RRULE can be generated with the given fields.
@@ -132,22 +166,26 @@ class EventForm(Form):
                 self.end_date.errors.append(message)
                 result = False
 
-        if self.weekly.data and self.start_date.data:
-            weekday = WEEKDAYS[self.start_date.data.weekday()][0]
-            if weekday not in self.weekly.data:
-                message = _("The weekday of the start date must be selected.")
-                self.weekly.errors.append(message)
+        if self.repeat.data == 'weekly':
+            if self.weekly.data and self.start_date.data:
+                weekday = WEEKDAYS[self.start_date.data.weekday()][0]
+                if weekday not in self.weekly.data:
+                    message = _(
+                        "The weekday of the start date must be selected.")
+                    self.weekly.errors.append(message)
+                    result = False
+
+            if self.weekly.data and not self.end_date.data:
+                message = _(
+                    "Please set and end date if the event is recurring.")
+                self.end_date.errors.append(message)
                 result = False
 
-        if self.weekly.data and not self.end_date.data:
-            message = _("Please set and end date if the event is recurring.")
-            self.end_date.errors.append(message)
-            result = False
-
-        if self.end_date.data and not self.weekly.data:
-            message = _("Please select a weekday if the event is recurring.")
-            self.weekly.errors.append(message)
-            result = False
+            if self.end_date.data and not self.weekly.data:
+                message = _(
+                    "Please select a weekday if the event is recurring.")
+                self.weekly.errors.append(message)
+                result = False
 
         return result
 
@@ -195,24 +233,34 @@ class EventForm(Form):
             model.timezone
         )
 
-        if self.weekly.data and self.end_date.data:
-            until_date = to_timezone(
-                replace_timezone(
-                    datetime(
-                        self.end_date.data.year,
-                        self.end_date.data.month,
-                        self.end_date.data.day,
-                        self.end_time.data.hour,
-                        self.end_time.data.minute
-                    ), model.timezone
-                ), 'UTC'
-            )
-            model.recurrence = (
-                "RRULE:FREQ=WEEKLY;WKST=MO;BYDAY={0};UNTIL={1}".format(
-                    ','.join(self.weekly.data),
-                    until_date.strftime('%Y%m%dT%H%M%SZ')
+        if self.repeat.data == 'without':
+            self.recurrence = None
+        elif self.repeat.data == 'weekly':
+            if self.weekly.data and self.end_date.data:
+                until_date = to_timezone(
+                    replace_timezone(
+                        datetime(
+                            self.end_date.data.year,
+                            self.end_date.data.month,
+                            self.end_date.data.day,
+                            self.end_time.data.hour,
+                            self.end_time.data.minute
+                        ), model.timezone
+                    ), 'UTC'
                 )
+                model.recurrence = (
+                    "RRULE:FREQ=WEEKLY;WKST=MO;BYDAY={0};UNTIL={1}".format(
+                        ','.join(self.weekly.data),
+                        until_date.strftime('%Y%m%dT%H%M%SZ')
+                    )
+                )
+        elif self.repeat.data == 'dates':
+            model.recurrence = '\n'.join(
+                f"RDATE:{d.strftime('%Y%m%dT000000Z')}"
+                for d in self.parsed_dates
             )
+        else:
+            raise NotImplementedError
 
     def process_obj(self, model):
         """ Stores the page values on the form. """
@@ -228,10 +276,48 @@ class EventForm(Form):
             self.end_date.data = last_occurrence.date()
 
             rule = rrulestr(model.recurrence)
-            if rule._freq == rrule.WEEKLY:
+
+            if getattr(rule, '_freq', None) == rrule.WEEKLY:
+                self.repeat.data = 'weekly'
+                self.dates.data = self.dates_to_json(None)
                 self.weekly.data = [
                     WEEKDAYS[day][0] for day in rule._byweekday
                 ]
+            else:
+                self.repeat.data = 'dates'
+                self.dates.data = self.dates_to_json(tuple(rule))
+        else:
+            self.dates.data = self.dates_to_json(None)
+            self.repeat.data = 'without'
 
         if model.meta:
             self.email.data = model.meta.get('submitter_email')
+
+    @cached_property
+    def parsed_dates(self):
+        return self.json_to_dates(self.dates.data)
+
+    def json_to_dates(self, text=None):
+        result = []
+
+        for value in json.loads(text or '{}').get('values', []):
+            result.append(date(*map(int, value['date'].split('-'))))
+
+        return result
+
+    def dates_to_json(self, dates=None):
+        dates = dates or []
+
+        return json.dumps({
+            'labels': {
+                'date': self.request.translate(_("Date")),
+                'add': self.request.translate(_("Add")),
+                'remove': self.request.translate(_("Remove")),
+            },
+            'values': [
+                {
+                    'date': d.strftime('%Y-%m-%d'),
+                    'error': self.date_errors.get(ix, "")
+                } for ix, d in enumerate(dates)
+            ]
+        })